requirements.txt
django
RPi.GPIO
pi-rc522
